Move the component files related to Retrigger to a subdirectory
This also moves the strings to the Strings file
